requirements.txt
torch~=2.5.1
timm~=1.0.11
torchvision~=0.20.1
pillow~=11.0.0
einops~=0.8.0
matplotlib~=3.9.2